+++ Components/BuiltInPalettes/init.lua
@@ -5,6 +5,7 @@ local Roact = require(includes:FindFirstChild("Roact"))
 
 local Components = root:FindFirstChild("Components")
 local ColorBrewerPalettes = require(Components:FindFirstChild("ColorBrewerPalettes"))
+local ColorVariations = require(Components:FindFirstChild("ColorVariations"))
 local Palette = require(Components:FindFirstChild("Palette"))
 
 local BrickColors = require(script:FindFirstChild("BrickColors"))
@@ -29,6 +30,14 @@ return {
 
         getContent = function()
             return Roact.createElement(ColorBrewerPalettes)
+        end
+    },
+
+    {
+        name = "Variations",
+        
+        getContent = function()
+            return Roact.createElement(ColorVariations)
         end
     },